Start consolidating xxxFromStore
