+++ A3A/addons/gui/dialogues/dialogs.hpp
@@ -13,8 +13,5 @@ Maintainer: DoomMetal
 #include "arsenalLimitsDialog.hpp"
 #include "setupDialog.hpp"
 #include "teamLeaderBuilder.hpp"
-#include "AAFromStore.hpp"
-#include "ATFromStore.hpp"
-#include "MGFromStore.hpp"
-#include "roadblockFromStore.hpp"
 #include "vehicleFromGarage.hpp"
+#include "vehicleFromStore.hpp"
